feat: onboarding 2
