+++ src/components/organisms/dashboard/templates/index.ts
@@ -1,2 +1,2 @@
-export { ProjectTemplatesSection } from "@components/organisms/dashboard/templates/section";
+export { TemplatesCatalog } from "@components/organisms/dashboard/templates/catalog";
 export { ProjectTemplatesTabs } from "@components/organisms/dashboard/templates/tabs";
