+++ src/store/useTemplatesStore.ts
@@ -13,14 +13,7 @@ import {
 } from "@constants";
 import { LoggerService, TemplateStorageService } from "@services";
 import { StoreName } from "@src/enums";
-import {
-	GitHubCommit,
-	TemplateCardWithFiles,
-	TemplateCategory,
-	TemplateMetadataWithCategory,
-	TemplateState,
-} from "@src/interfaces/store";
-import { fetchAndUnpackZip, processReadmeFiles } from "@utilities";
+import { GitHubCommit, TemplateCategory, TemplateState } from "@src/interfaces/store";
 
 const sortCategories = (categories: TemplateCategory[], order: string[]) => {
 	return categories.sort((a, b) => {
@@ -68,59 +61,6 @@ const store = (set: any, get: any): TemplateState => ({
 		});
 
 		set({ isLoading: true, error: null });
-
-		const processTemplates = async (
-			zipUrl: string
-		): Promise<{
-			categories?: TemplateCategory[];
-			error?: string;
-			templateMap?: Record<string, TemplateMetadataWithCategory>;
-		}> => {
-			const processTemplateCard = async (cardWithFiles: TemplateCardWithFiles, categoryName: string) => {
-				const { getTemplateStorage } = get();
-				const tmpStorage = getTemplateStorage();
-				await tmpStorage.storeTemplateFiles(cardWithFiles.assetDirectory, cardWithFiles.files);
-
-				return {
-					assetDirectory: cardWithFiles.assetDirectory,
-					title: cardWithFiles.title,
-					description: cardWithFiles.description,
-					integrations: cardWithFiles.integrations,
-					filesIndex: Object.keys(cardWithFiles.files),
-					category: categoryName,
-				};
-			};
-
-			const result = await fetchAndUnpackZip(zipUrl);
-			if (!("structure" in result) || result.error) {
-				return { error: couldntFetchTemplates };
-			}
-
-			const processedCategories = processReadmeFiles(result.structure);
-			const templateMap: Record<string, TemplateMetadataWithCategory> = {};
-
-			await Promise.all(
-				processedCategories.map(async ({ cards, name }) => {
-					const processedCards = await Promise.all(cards.map((card) => processTemplateCard(card, name)));
-					processedCards.forEach((cardData) => {
-						templateMap[cardData.assetDirectory] = cardData;
-					});
-				})
-			);
-
-			const categories = Object.values(templateMap).reduce((acc, template) => {
-				const category = acc.find((c) => c.name === template.category);
-				if (category) {
-					category.templates.push(template);
-				} else {
-					acc.push({ name: template.category, templates: [template] });
-				}
-
-				return acc;
-			}, [] as TemplateCategory[]);
-
-			return { templateMap, categories };
-		};
 
 		try {
 			let shouldFetchTemplates = !Object.keys(get().templateMap).length;
